+++ src/main/php/lang/mirrors/Constant.class.php
@@ -1,5 +1,10 @@
 <?php namespace lang\mirrors;
 
+/**
+ * A class constant
+ *
+ * @test   xp://lang.mirrors.unittest.ConstantTest
+ */
 class Constant extends \lang\Object {
   private $name, $value;
 
@@ -13,4 +18,18 @@ class Constant extends \lang\Object {
 
   /** @return var */
   public function value() { return $this->value; }
+
+  /**
+   * Creates a string representation
+   *
+   * @return string
+   */
+  public function toString() {
+    return $this->getClassName().'('.$this.')';
+  }
+
+  /** @return string */
+  public function __toString() {
+    return $this->name.'= '.var_export($this->value, true);
+  }
 }