+++ src/test/php/lang/mirrors/unittest/ConstantTest.class.php
@@ -27,11 +27,10 @@ class ConstantTest extends \unittest\TestCase {
     $this->assertEquals($this->type, (new Constant($this->type, 'fixture', 'value'))->declaredIn());
   }
 
-
   #[@test]
   public function string_representation() {
     $this->assertEquals(
-      'lang.mirrors.Constant(fixture= \'value\')',
+      'lang.mirrors.Constant(fixture= "value")',
       (new Constant($this->type, 'fixture', 'value'))->toString()
     );
   }